requirements.txt
asgiref==3.5.0
distlib==0.3.4
Django==4.0.2
django-cors-headers==3.12.0
django-model-utils==4.2.0
django-simple-history==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.6.0
mysql-connector-python==8.0.29
mysqlclient==2.1.0
Pillow==9.1.0
platformdirs==2.5.0
protobuf==3.20.1
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.14.1
